+++ CHANGELOG.md
@@ -33,6 +33,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Residue of Type UNK to residue class and templates.
   - 3 letter name: UNK, 1 letter name: X, full name Unknown
   - Template Coordinates and bonds: same as Alanine
+- applyTransformation() methods to Protein, PolypeptideChain, Residue, and Atom
+  - applyTransformation() naively applies the transformation to all components
+    of that class. Eg, when called on Protein, it calls the applyTransformation
+    for each PolypeptideChain, which does the same for every Residue, which then
+    applies the transformation to each Atom - whose coordinates are then
+    transformed.
+- getOmegaAngle(), getPhiAngle(), and getPsiAngle() in PolypeptideChain. Each
+  takes a residueID and returns the angle if it can be calculated or 1000
+  otherwise.
+  - a test to TestProtein to check that all the angles for 1rop are calculated
+    correctly. 
 ### Changed
 - Moved SequenceAligner into new package sequence
 - Protein and PolypeptideChain getSequence methods to use new sequence package